requirement.txt
fastapi
uvicorn
numpy
pandas
sqlalchemy
databases
asyncpg
psycopg2
python-multipart  
pydantic          
httpx             
colorlog
python-dotenv
requests
apscheduler
joblib
lightgbm
sklearn-pandas
mysql-connector-python
PyMySQL
mysqlclient
tqdm